fix: wrap client routes in suspense
